Type dashboards/graphics/graph & move cell types to types/dashboard
